chore: organize certs-v5 unit tests
