+++ packages/certs-v5/test/unit/commands/certs/auto/disable.unit.test.js
@@ -7,7 +7,7 @@ chai.use(chaiAsPromised)
 
 let expect = chai.expect
 let nock = require('nock')
-let certs = require('../../../../commands/certs/auto/disable.js')
+let certs = require('../../../../../commands/certs/auto/disable.js')
 
 describe('heroku certs:auto:disable', function () {
   beforeEach(function () {
