+++ packages/certs-v5/test/unit/commands/certs/auto/refresh.unit.test.js
@@ -3,7 +3,7 @@
 
 let expect = require('chai').expect
 let nock = require('nock')
-let certs = require('../../../../commands/certs/auto/refresh.js')
+let certs = require('../../../../../commands/certs/auto/refresh.js')
 
 describe('heroku certs:auto:enable', function () {
   beforeEach(function () {
